[nasa/nos3#378] Split device into a new child file set;
