+++ fsw/cfs/platform_inc/cam_platform_cfg.h
@@ -29,39 +29,23 @@ ivv-itc@lists.nasa.gov
 #define _CAM_PLATFORM_CFG_H_
 
 /*
-** CAM Child Task Definitions
+** Default CAM Configuration
 */
-#define CAM_CHILD_TASK_NAME              	"CAM_CHILD_TASK"
-#define CAM_CHILD_TASK_STACK_SIZE       	2048
-#define CAM_CHILD_TASK_PRIORITY          	205
-#define CAM_RUN						 	    0
-#define CAM_PAUSE						    1
-#define CAM_STOP						    2
-#define CAM_TIME                            3
-#define CAM_LOW_VOLTAGE                     4
-
-#define CAM_MUTEX_NAME                      "CAM_MUTEX"
-#define CAM_SEM_NAME                        "CAM_SEM"
-
-/****************************************************/
-/* Sensor related definition 						*/
-/****************************************************/
-#define BMP 	0
-#define JPEG	1
-
-#define OV2640_160x120 		0	//160x120
-#define OV2640_176x144 		1	//176x144
-#define OV2640_320x240 		2	//320x240
-#define OV2640_352x288 		3	//352x288
-#define OV2640_640x480		4	//640x480
-#define OV2640_800x600 		5	//800x600
-#define OV2640_1024x768		6	//1024x768
-#define OV2640_1280x1024	7	//1280x1024
-#define OV2640_1600x1200	8	//1600x1200
-
-/****************************************************/
-/* ArduChip related definition 						*/
-/****************************************************/
-#define ARDUCHIP_MODE      		0x02  //Mode register
+#ifndef CAM_CFG
+    #define CAM_I2C_BUS                         2
+    #define CAM_SPEED                           1000000
+    #define CAP_DONE_MASK                       0x08
+    #define CAM_TIMEOUT                         100
+    #define CAM_DATA_SIZE                       1010
+    #define CAM_CHILD_TASK_NAME              	"CAM_CHILD_TASK"
+    #define CAM_CHILD_TASK_STACK_SIZE       	2048
+    #define CAM_CHILD_TASK_PRIORITY          	205
+    #define CAM_MUTEX_NAME                      "CAM_MUTEX"
+    #define CAM_SEM_NAME                        "CAM_SEM"
+    // Select Hardware (only 1)
+    //#define OV2640
+    #define OV5640 
+    //#define OV5642
+#endif 
 
 #endif
